requirements.txt
Flask==2.0.1
Flask-Login==0.5.0
Flask-SQLAlchemy==2.5.1
APScheduler==3.9.1
pywin32==307
python-vlc==3.0.12118
requests==2.26.0
SQLAlchemy==1.4.32
Werkzeug==2.0.1
selenium==4.10.0